Create profile page , getUsePostsController, GetUserPostsQuery , from Clint , server and database Relates # 48
