+++ public/html/profile.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="../css/profile.css">
+  <link rel="stylesheet" href="/css/profile.css">
   <title>Document</title>
 </head>
 <body>
@@ -46,7 +46,7 @@
         <div id="profile-pic">
           <img src="https://imagizer.imageshack.com/img921/3072/rqkhIb.jpg">
         </div>
-        <div id="u-name">Himalaya Singh</div>
+        <div id="userName">-</div>
         <div class="tb" id="m-btns">
           <div class="td">
             <div class="m-btn"><i class="material-icons">format_list_bulleted</i><span>Activity log</span></div>
@@ -150,5 +150,7 @@
       </div>
     </div>
   </main>
+  <script src="/js/createPostElement.js"></script>
+  <script src="/js/profile.js"></script>
 </body>
 </html>
