+++ public/js/login.js
@@ -1,6 +1,9 @@
 const btn = document.getElementById("btn");
 const divMessage = document.getElementById('divMessage')
-
+const homeBtn =  document.getElementById('home')
+// homeBtn.addEventListener('click', () => {
+//   window.location.href = "/profile.html";
+// })
 btn.addEventListener("click", (e) => {
   e.preventDefault();
   const email = document.getElementById("email");
