+++ server/controllers/index.js
@@ -2,5 +2,5 @@ const addLoginController = require("./middlewares/loginController");
 const signupContoller = require("./middlewares/signupController");
 const getPostsController = require("./middlewares/postsController");
 const getusername = require("./middlewares/usernameController");
-
-module.exports = { addLoginController, signupContoller, getPostsController , getusername};
+const getUserPostsController = require("./middlewares/getuserPostController");
+module.exports = { addLoginController, signupContoller, getPostsController , getusername, getUserPostsController};
