+++ server/database/index.js
@@ -2,5 +2,6 @@ const getPostsQuery = require("./queries/getPostsQuery");
 const signupQuery = require("./queries/signupQuery");
 const Build = require("./config/build");
 const getusernameQuery = require("./queries/getusernameQuery");
+const getUserPostsQuery = require("./queries/getuserPostQuery");
 
-module.exports = { signupQuery, Build, getPostsQuery,getusernameQuery };
+module.exports = { signupQuery, Build, getPostsQuery,getusernameQuery,getUserPostsQuery };
