+++ server/database/queries/getusernameQuery.js
@@ -2,7 +2,7 @@ const connection = require("../connection")
 
 const getusernameQuery = (id) => {
     const sql = {
-        text: 'select firstname, surname from users where id=$1',
+        text: 'select firstname, surname, id from users where id=$1',
         values:[id]
     }
     return connection.query(sql);
