+++ server/routes/index.js
@@ -5,6 +5,7 @@ const {
   signupContoller,
   getPostsController,
   getusername,
+  getUserPostsController,
 } = require("../controllers");
 const { clientError, serverError } = require("../controllers/errors/errors");
 const homePage = require("../controllers/middlewares/homeController");
@@ -13,6 +14,7 @@ router.post("/login", addLoginController);
 router.post("/signup", signupContoller);
 router.get("/posts", getPostsController);
 router.get('/username', getusername)
+router.get('/userPosts', getUserPostsController)
 router.use(homePage);
 router.use(clientError);
 router.use(serverError);
